+++ src/state/src/actor/states/actor_pursuit_state.cpp
@@ -12,7 +12,7 @@ void ActorPursuitState::Enter(Actor * actor) {}
 
 std::unique_ptr<ActorState> ActorPursuitState::Update(
 	Actor * actor,
-	int64_t delta_time
+	float delta_time
 ) {
 	auto target = actor->GetAttackTarget();
 	if (actor->GetHp() <= 0) {
